+++ source/egov-erp/egov-eis/src/main/java/org/egov/pims/service/EmployeeService.java
@@ -71,8 +71,8 @@ public interface EmployeeService
 	public abstract List getListOfPersonalInformationByEmpIdsList(List empIdsList);
 	public List getListOfEmployeeWithoutAssignment(Date fromdate);/*new*/
 	public Assignment getLatestAssignmentForEmployeeByToDate(Integer empId,Date todate) throws Exception;
-	public PersonalInformation getEmployee(Integer deptId, Integer designationId, Integer Boundaryid)throws TooManyValuesException, NoSuchObjectException;
-	public PersonalInformation getEmployeeByFunctionary(Integer deptId, Integer designationId, Integer Boundaryid,Integer functionaryId)throws TooManyValuesException, NoSuchObjectException;
+	public PersonalInformation getEmployee(Integer deptId, Integer designationId, Long Boundaryid)throws TooManyValuesException, NoSuchObjectException;
+	public PersonalInformation getEmployeeByFunctionary(Integer deptId, Integer designationId, Long Boundaryid,Integer functionaryId)throws TooManyValuesException, NoSuchObjectException;
 	public Assignment getLastAssignmentByEmp(Integer empId);
 
 	//to delete grade master
